requirements.txt
Flask==2.0.3
requests==2.26.0
requests-oauthlib==1.3.0
python-dotenv==0.20.0
